backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
redis==5.2.1
SQLAlchemy==2.0.36
alembic==1.13.3
openai==1.57.0
pydantic==2.9.2

